final version after presentation

+++ SilencerStoreProject/src/com/silencer/controller/frontend/shoppingcart/ShoppingCart.java
@@ -52,8 +52,8 @@ public class ShoppingCart {
 		return total;
 	}
 	
-	public double getTotalAmount() {
-		double total = 0.0f;
+	public float getTotalAmount() {
+		float total = 0.0f;
 		
 		Iterator<Silencer> iterator = cart.keySet().iterator();
